+++ changelog.md
@@ -5,6 +5,9 @@
 - Bug fixes
 - Remote server/client improvements
 - Better support for some CBZ files (if SUMMARY or URL files are present in it)
+- Fix cover images not deleted on story delete
+- Fix some images not supported because not jpeg-able (now try again in png)
+- Fix some covers not found (normal and remote hopefully)
 
 ## Version 1.6.2